refactor(nicknamer): move tests to `mod.rs` and update accessibility
Relocated the `reveal` tests from `commands/reveal.rs` to `mod.rs` for better organization. Updated trait and struct access modifiers in `Nicknamer` to support test visibility and ensure compatibility.
